make ticket storage work throughout all booking stages
