+++ src/components/BookingSummary.tsx
@@ -6,58 +6,54 @@ import { useState } from 'react';
 import clsx from 'clsx';
 import { cn } from '../lib/utils';
 import { SummaryRow } from '../types/booking';
-import { SeatType } from '../types/booking';
+import { Link } from 'react-router-dom';
+import { useTicketStore } from '../stores/ticket';
 
 // This logic is purely made up for now
-function returnSeatType(seatCode: string) {
-  const row = seatCode.split('-')[0];
-  switch (row) {
-    case 'A':
-      return 'Front';
-    // assume H is the last row
-    case 'H':
-      return 'Back';
-    default:
-      return 'Middle';
-  }
-}
-
-// This is purely made up for now
-const seatPrices = {
-  Front: 12.99,
-  Middle: 14.75,
-  Back: 16.99,
-};
+// function returnSeatType(seatCode: string) {
+//   const row = seatCode.split('-')[0];
+//   switch (row) {
+//     case 'A':
+//       return 'Front';
+//     // assume H is the last row
+//     case 'H':
+//       return 'Back';
+//     default:
+//       return 'Middle';
+//   }
+// }
 
 // Transform selected seats into summaries
-function createBookingSummary(selectedSeats: SeatType[]) {
-  const summaries: SummaryRow[] = [];
-  selectedSeats.forEach(seat => {
-    const type = returnSeatType(seat.code);
-    if (!summaries.some(row => row.type === type)) {
-      summaries.push({ type: type, amount: 1, price: seatPrices[type] });
-    } else {
-      summaries.forEach(row => {
-        if (row.type === type) {
-          row.amount += 1;
-        }
-      });
-    }
-  });
-  return summaries;
-}
+// function createBookingSummary(selectedSeats: SeatType[]) {
+//   const summaries: SummaryRow[] = [];
+//   selectedSeats.forEach(seat => {
+//     const type = returnSeatType(seat.code);
+//     if (!summaries.some(row => row.type === type)) {
+//       summaries.push({ type: type, amount: 1, price: seatPrices[type] });
+//     } else {
+//       summaries.forEach(row => {
+//         if (row.type === type) {
+//           row.amount += 1;
+//         }
+//       });
+//     }
+//   });
+//   return summaries;
+// }
+
+type BookingSummaryProps = {
+  buttonLink: string;
+  summaries: SummaryRow[];
+};
 
 export function BookingSummary({
-  selectedSeats,
-}: {
-  selectedSeats: SeatType[];
-}) {
+  summaries,
+  buttonLink,
+}: BookingSummaryProps) {
   const [isOpen, setIsOpen] = useState(true);
   const toggleSlide = () => {
     setIsOpen(!isOpen);
   };
-
-  const summaries = createBookingSummary(selectedSeats);
 
   return (
     <div className="w-full bg-dark-light rounded-3xl px-5 pt-7 pb-6">
@@ -87,7 +83,7 @@ export function BookingSummary({
         <div
           className={cn(
             'border border-t-[1px] border-white-dimmed-heavy border-b-0 mt-4',
-            selectedSeats.length === 0 && 'border-t-0'
+            summaries.length === 0 && 'border-t-0'
           )}
         ></div>
       </Transition>
@@ -97,12 +93,12 @@ export function BookingSummary({
             Total Price
           </span>
           <span className="text-xl font-700 text-white">
-            {`$${summaries
-              .reduce((acc, row) => acc + row.amount * row.price, 0)
-              .toFixed(2)}`}
+            {`$${summaries.reduce((acc, row) => acc + row.amount * row.price, 0).toFixed(2)}`}
           </span>
         </div>
-        <Button className='w-3/5'>Book Ticket</Button>
+        <Link className="w-3/5" to={buttonLink}>
+          <Button className="w-full">Book Ticket</Button>
+        </Link>
       </div>
     </div>
   );
