+++ src/stores/ticket.ts
@@ -1,5 +1,6 @@
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
+import { SeatType } from '../types/booking';
 
 type TicketType = {
   id: string;
